Rename splitting components
