+++ client/views/root/MainLayout/MainLayout.tsx
@@ -1,138 +1,17 @@
-/* eslint-disable react/no-multi-comp */
-import React, { lazy, ReactElement, useCallback } from 'react';
+import React, { ReactElement, useMemo } from 'react';
 
-import { Roles, Users } from '../../../../app/models/client';
-import { IUser } from '../../../../definition/IUser';
-import { useLayout } from '../../../contexts/LayoutContext';
-import { useCurrentRoute, useRoutePath } from '../../../contexts/RouterContext';
-import { useSetting } from '../../../contexts/SettingsContext';
-import { useUser, useUserId } from '../../../contexts/UserContext';
-import { useReactiveValue } from '../../../hooks/useReactiveValue';
 import BlazeTemplate from '../BlazeTemplate';
-import PageLoading from '../PageLoading';
-import { useAllowRead } from './useAllowRead';
-import { useCollectionsAvailability } from './useCollectionsAvailability';
-import { useCustomScript } from './useCustomScript';
-import { useIframeLogin } from './useIframeLogin';
-import { useViewportScrolling } from './useViewportScrolling';
-
-const MainLayout1 = (): ReactElement => {
-	const iframeLoginUrl = useIframeLogin();
-
-	if (iframeLoginUrl) {
-		return <iframe src={iframeLoginUrl} style={{ height: '100%', width: '100%' }} />;
-	}
-
-	return <BlazeTemplate template='loginLayout' />;
-};
-
-const ResetPasswordPage = lazy(() => import('../../login/ResetPassword/ResetPassword'));
-const AccountSecurityPage = lazy(() => import('../../account/security/AccountSecurityPage'));
-
-const MainLayout5 = ({ center }: MainLayoutProps): ReactElement => {
-	const { isEmbedded: embeddedLayout } = useLayout();
-	const [currentRouteName = '', currentParameters = {}] = useCurrentRoute();
-	const currentRoutePath = useRoutePath(currentRouteName, currentParameters);
-	const removeSidenav = useReactiveValue(
-		useCallback(() => embeddedLayout && !currentRoutePath?.startsWith('/admin'), [currentRoutePath, embeddedLayout]),
-	);
-	const readReceiptsEnabled = useSetting('Message_Read_Receipt_Store_Users');
-
-	return (
-		<div id='rocket-chat' className={[embeddedLayout ? 'embedded-view' : undefined, 'menu-nav'].filter(Boolean).join(' ')}>
-			{!removeSidenav ? <BlazeTemplate template='sideNav' /> : null}
-			<div
-				className={['rc-old', 'main-content', 'content-background-color', readReceiptsEnabled ? 'read-receipts-enabled' : undefined]
-					.filter(Boolean)
-					.join(' ')}
-			>
-				{center ? <BlazeTemplate template={center} /> : null}
-			</div>
-		</div>
-	);
-};
-
-const MainLayout4 = ({ center }: MainLayoutProps): ReactElement => {
-	const { isEmbedded: embeddedLayout } = useLayout();
-	const user = useUser() as IUser | null;
-	const tfaEnabled = useSetting('Accounts_TwoFactorAuthentication_Enabled');
-	const require2faSetup = useReactiveValue(
-		useCallback(() => {
-			// User is already using 2fa
-			if (!user || user?.services?.totp?.enabled || user?.services?.email2fa?.enabled) {
-				return false;
-			}
-
-			const mandatoryRole = Roles.findOne({ _id: { $in: user.roles }, mandatory2fa: true });
-			return mandatoryRole !== undefined && tfaEnabled;
-		}, [tfaEnabled, user]),
-	);
-
-	if (require2faSetup) {
-		return (
-			<main id='rocket-chat' className={embeddedLayout ? 'embedded-view' : undefined}>
-				<div className='rc-old main-content content-background-color'>
-					<AccountSecurityPage />
-				</div>
-			</main>
-		);
-	}
-
-	return <MainLayout5 center={center} />;
-};
-
-const MainLayout3 = ({ center }: MainLayoutProps): ReactElement => {
-	const requirePasswordChange = (useUser() as IUser | null)?.requirePasswordChange === true;
-
-	if (requirePasswordChange) {
-		return <ResetPasswordPage />;
-	}
-
-	return <MainLayout4 center={center} />;
-};
-
-const MainLayout2 = ({ center }: MainLayoutProps): ReactElement => {
-	useViewportScrolling();
-	useCustomScript();
-
-	const uid = useUserId();
-	const allowAnonymousRead = useSetting('Accounts_AllowAnonymousRead');
-
-	const hasUsername = useReactiveValue(
-		useCallback(() => {
-			if (!uid) {
-				return allowAnonymousRead;
-			}
-
-			const user = uid ? (Users.findOneById(uid, { fields: { username: 1 } }) as IUser | null) : null;
-			return user?.username ?? false;
-		}, [uid, allowAnonymousRead]),
-	);
-
-	if (!hasUsername) {
-		return <BlazeTemplate template='username' data={{ center }} />;
-	}
-
-	return <MainLayout3 center={center} />;
-};
+import AuthenticationCheck from './AuthenticationCheck';
+import Preload from './Preload';
 
 type MainLayoutProps = {
 	center?: string;
 };
 
-const MainLayout = ({ center }: MainLayoutProps): ReactElement => {
-	const ready = useCollectionsAvailability();
-	const allowedRead = useAllowRead(ready);
-
-	if (!ready) {
-		return <PageLoading />;
-	}
-
-	if (!allowedRead) {
-		return <MainLayout1 />;
-	}
-
-	return <MainLayout2 center={center} />;
-};
+const MainLayout = ({ center }: MainLayoutProps): ReactElement => (
+	<Preload>
+		<AuthenticationCheck>{useMemo(() => (center ? <BlazeTemplate template={center} /> : null), [center])}</AuthenticationCheck>
+	</Preload>
+);
 
 export default MainLayout;
